Map Implementation
Implementation of maps, tiles or textures, and code to read in new maps
and tiles for the future. Approach uses a Two-Dimentional Array to draw
out the maps with properly arranged .TXT files that make the creation of
maps more efficient. It also uses number identification for each tile or
texture included.'Space for tiles can always be incremented'.