Collision implemented, camera implemented
The collision mechanism was implemented, right now the only tile that
has collision is the wall tile. I had to create a "hit-box" for the
player to make this collision happen.
CollisionChecker.java is the one in charge of checking possible
collision.
A bigger map of 50x50 has been implemented for testing the camera that
follows the player. The player stays at the center the whole time. The
map will load as the player walks, this to improve performance.